+++ src/auth/auth.module.ts
@@ -7,16 +7,16 @@ import { RbacJwtStrategy } from "src/common/guards/rbac.strategy";
 import { UserAdapter } from "../user/useradapter";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { User } from "../user/entities/user-entity";
-import { FieldValues } from "../user/entities/field-value-entities";
-import { Field } from "src/user/entities/field-entity";
+import { FieldValues } from "src/fields/entities/fields-values.entity";
+import { Fields } from "src/fields/entities/fields.entity";
 import { CohortMembers } from "src/cohortMembers/entities/cohort-member.entity";
 import { KeycloakService } from "src/common/utils/keycloak.service";
 import { HasuraModule } from "src/adapters/hasura/hasura.module";
-import { PostgresModule } from "src/adapters/postgres/potsgres-module";
+import { PostgresModule } from "src/adapters/postgres/postgres-module";
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([User, FieldValues, Field, CohortMembers]),
+    TypeOrmModule.forFeature([User, FieldValues, Fields, CohortMembers]),
     HttpModule,
     HasuraModule,
     PostgresModule,
