+++ src/authRbac/authRbac.module.ts
@@ -5,7 +5,7 @@ import { AuthRbacService } from "./authRbac.service";
 import { AuthRbacController } from "./authRbac.controller";
 import { UserAdapter } from "src/user/useradapter";
 import { HasuraModule } from "src/adapters/hasura/hasura.module";
-import { PostgresModule } from "src/adapters/postgres/potsgres-module";
+import { PostgresModule } from "src/adapters/postgres/postgres-module";
 import { PostgresRoleService } from "src/adapters/postgres/rbac/role-adapter";
 import { Role } from "src/rbac/role/entities/role.entity";
 import { TypeOrmModule } from "@nestjs/typeorm";
