+++ src/user/user.module.ts
@@ -3,11 +3,11 @@ import { UserController } from "./user.controller";
 import { HttpModule } from "@nestjs/axios";
 import { UserAdapter } from "./useradapter";
 import { HasuraModule } from "src/adapters/hasura/hasura.module";
-import { PostgresModule } from "src/adapters/postgres/potsgres-module";
+import { PostgresModule } from "src/adapters/postgres/postgres-module";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { User } from "./entities/user-entity";
-import { FieldValues } from "./entities/field-value-entities";
-import { Field } from "./entities/field-entity";
+import { FieldValues } from "../fields/entities/fields-values.entity";
+import { Fields } from "src/fields/entities/fields.entity";
 import { CohortMembers } from "src/cohortMembers/entities/cohort-member.entity";
 import { UserTenantMapping } from "src/userTenantMapping/entities/user-tenant-mapping.entity";
 import { Tenants } from "src/userTenantMapping/entities/tenant.entity";
@@ -18,7 +18,7 @@ import { Role } from "src/rbac/role/entities/role.entity";
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([User, FieldValues, Field, CohortMembers,UserTenantMapping,Tenants,UserRoleMapping,Cohort,Role]),
+    TypeOrmModule.forFeature([User, FieldValues, Fields, CohortMembers,UserTenantMapping,Tenants,UserRoleMapping,Cohort,Role]),
     HttpModule,
     HasuraModule,
     PostgresModule,
